+++ synapse/lib/stormlib/task.py
@@ -1,4 +1,3 @@
-import asyncio
 import logging
 
 import synapse.exc as s_exc