+++ synapse/tests/test_model_inet.py
@@ -2314,6 +2314,16 @@ class InetModelTest(s_t_utils.SynTest):
             self.len(1, nodes)
             self.eq(nodes[0].ndef, ('inet:whois:email', ('woot.com', 'pennywise@vertex.link')))
 
+            with self.getLoggerStream('synapse.datamodel') as stream:
+                nodes = await core.nodes('[ inet:whois:record=* :text="Contact: pennywise@vertex.link" ]')
+                self.len(1, nodes)
+                self.eq(nodes[0].get('text'), 'contact: pennywise@vertex.link')
+                self.none(nodes[0].get('fqdn'))
+
+            stream.seek(0)
+            data = stream.read()
+            self.notin('onset() error for inet:whois:record:text', data)
+
     async def test_whois_iprecord(self):
         async with self.getTestCore() as core:
             contact = s_common.guid()
@@ -2688,11 +2698,18 @@ class InetModelTest(s_t_utils.SynTest):
 
             q = '''
             [ inet:service:platform=(slack,)
+                :id=foo
                 :url="https://slack.com"
                 :urls=(https://slacker.com,)
+                :zones=(slack.com, slacker.com)
                 :name=Slack
                 :names=("slack chat",)
                 :desc=' Slack is a team communication platform.\n\n Be less busy.'
+                :parent={[ inet:service:platform=({"name": "salesforce"}) ]}
+                :status=available
+                :period=(2022, 2023)
+                :creator={[ inet:service:account=({"id": "bar"}) ]}
+                :remover={[ inet:service:account=({"id": "baz"}) ]}
                 :provider={ ou:org:name=$provname }
                 :provider:name=$provname
             ]
@@ -2700,19 +2717,35 @@ class InetModelTest(s_t_utils.SynTest):
             nodes = await core.nodes(q, opts=opts)
             self.len(1, nodes)
             self.eq(nodes[0].ndef, ('inet:service:platform', s_common.guid(('slack',))))
+            self.eq('foo', nodes[0].get('id'))
             self.eq(nodes[0].get('url'), 'https://slack.com')
             self.eq(nodes[0].get('urls'), ('https://slacker.com',))
+            self.eq(nodes[0].get('zones'), ('slack.com', 'slacker.com'))
             self.eq(nodes[0].get('name'), 'slack')
             self.eq(nodes[0].get('names'), ('slack chat',))
             self.eq(nodes[0].get('desc'), ' Slack is a team communication platform.\n\n Be less busy.')
+            self.eq(nodes[0].repr('status'), 'available')
+            self.eq(nodes[0].repr('period'), ('2022-01-01T00:00:00Z', '2023-01-01T00:00:00Z'))
             self.eq(nodes[0].get('provider'), provider.ndef[1])
             self.eq(nodes[0].get('provider:name'), provname.lower())
             platform = nodes[0]
+
+            nodes = await core.nodes('inet:service:platform=(slack,) :parent -> *')
+            self.eq(['salesforce'], [n.get('name') for n in nodes])
+
+            nodes = await core.nodes('inet:service:platform=(slack,) :creator -> *')
+            self.eq(['bar'], [n.get('id') for n in nodes])
+
+            nodes = await core.nodes('inet:service:platform=(slack,) :remover -> *')
+            self.eq(['baz'], [n.get('id') for n in nodes])
 
             nodes = await core.nodes('[ inet:service:platform=({"name": "slack chat"}) ]')
             self.eq(nodes[0].ndef, platform.ndef)
 
             nodes = await core.nodes('[ inet:service:platform=({"url": "https://slacker.com"}) ]')
+            self.eq(nodes[0].ndef, platform.ndef)
+
+            nodes = await core.nodes('[ inet:service:platform=({"zone": "slacker.com"}) ]')
             self.eq(nodes[0].ndef, platform.ndef)
 
             q = '''